+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [v0.0.17] - 2022-MM-DD
 ### API Changes
+- `Chord::getStackedHeaps()` -> `Chord::getStackDataFrame()`
 
 ### New Features
-- `Interval Class`:
-    - `.isUnisson()`
+- `Interval Class`: 50 new classification methods
 
 ### Improved
 - `Interval Class`: